# swagger.yaml
swagger: "2.0"
info:
  version: "1.0.1"
  title: "Billing Account and Client API"
# during production,should point to your server machine
host: api.topcoder-dev.com
basePath: "/v3"
...
schemes:
- "https"
produces:
- application/json
consumes:
- application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
  /clients:
    get:
      tags:
        - client
      security:
        - Bearer: []
      operationId: findAllClients
      description: Retreive all clients. Availble to admin user only.
      parameters:
        - $ref: "#/parameters/offsetParam"
        - $ref: "#/parameters/limitParam"
        - name: filter
          required: true
          type: string
          in: query
          description: |
            Url encoded list of Supported filters
             - status
             - name
             - startDate
             - endDate
        - name: sort
          required: false
          description: |
            sort clients by status, name, startDate, endDate
          in: query
          type: string
      responses:
        '403':
          description: No permission or wrong token
          schema:
            $ref: "#/definitions/ErrorModel"
        '422':
          description: Invalid input
          schema:
            $ref: "#/definitions/ErrorModel"
        '200':
          description: A list of clients
          schema:
            $ref: "#/definitions/ClientListResponse"

    post:
      tags:
        - client
      security:
        - Bearer: []
      operationId: addNewClient
      description: Create a new client or update an existed client, Admin user only. If method is POST or not set, a new client will be created. You can update an exsiting client when method is PUT or PATCH
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/NewClientBodyParam'
      responses:
        '403':
          description: No permission or wrong token
          schema:
            $ref: "#/definitions/ErrorModel"
        '201':
          description: Returns the newly created client
          schema:
            $ref: "#/definitions/ClientResponse"
        '422':
          description: Invalid input
          schema:
            $ref: "#/definitions/ErrorModel"
  /clients/{clientId}:
    get:
      tags:
        - client
      security:
        - Bearer: []
      description: Retrieve client by id, Admin user only.
      responses:
        '403':
          description: No permission or wrong token
          schema:
            $ref: "#/definitions/ErrorModel"
        '404':
          description: Not found
          schema:
            $ref: "#/definitions/ErrorModel"
        '200':
          description: a client
          schema:
            $ref: "#/definitions/ClientResponse"
      parameters:
        - $ref: "#/parameters/clientIdParam"
      operationId: getClientById

    patch:
      tags:
        - client
      security:
        - Bearer: []
      operationId: updateClient
      description: Update a client, Admin user only.
      responses:
        '403':
          description: No permission or wrong token
          schema:
            $ref: "#/definitions/ErrorModel"
        '404':
          description: Not found
          schema:
            $ref: "#/definitions/ErrorModel"
        '200':
          description: Successfully updated client. Returns updated client
          schema:
            $ref: "#/definitions/ClientResponse"
        '422':
          description: Invalid input
          schema:
            $ref: "#/definitions/ErrorModel"
        default:
          description: error payload
          schema:
            $ref: '#/definitions/ErrorModel'
      parameters:
        - $ref: "#/parameters/clientIdParam"
        - name: body
          in: body
          required: true
          schema:
            $ref: "#/definitions/NewClientBodyParam"
...
parameters:
  clientIdParam:
    name: clientId
    in: path
    description: client identifier
    required: true
    type: integer
    format: int64
...
  offsetParam:
    name: offset
    description: "number of items to skip. Defaults to 0"
    in: query
    required: false
    type: integer
    format: int32
  limitParam:
    name: limit
    description: "max records to return. Defaults to 10"
    in: query
    required: false
    type: integer
    format: int32
definitions:
  ResponseMetadata:
    title: Metadata object for a response
    type: object
    properties:
      totalCount:
        type: integer
        format: int64
        description: Total count of the objects
  ErrorModel:
    type: object
    properties:
      id:
        type: string
        description: unique id identifying the request
      version:
        type: string
      result:
        type: object
        properties:
          success:
            type: boolean
          status:
            description: http status code
            type: integer
            format: int32
          debug:
            type: object
          content:
            type: object
  NewClient:
    type: object
    required:
       - name
       - status
    properties:
      name:
        type: string
        description: The client name (up to 64 characters)
      status:
        type: string
        description: the client status (ACTIVE or INACTIVE)
      startDate:
        type: string
        description: billing account start date, Datetime (GMT)
      endDate:
        type: string
        description: billing account end date, Datetime (GMT)
      codeName:
        type: string
        description: the client code name (optional)
      customerNumber:
        type: string
        description: the client customer number. This is an external reference code (optional)
...
  NewClientBodyParam:
    type: object
    properties:
      param:
        $ref: "#/definitions/NewClient"
      method:
        type: string
        description: Should be POST, PUT, or PATCH
...
  Client:
    type: object
    properties:
      id:
        description: unique identifier
        type: integer
        format: int64
      name:
        type: string
        description: The client name
      status:
        type: string
        description: The client status
      startDate:
        type: string
        description: billing account start date, Datetime (GMT)
      endDate:
        type: string
        description: billing account end date, Datetime (GMT)
      codeName:
        type: string
        description: The client code name
      customerNumber:
        type: string
        description: The client customer number
  ClientResponse:
    title: Single client
    type: object
    properties:
      id:
        type: string
        description: unique id identifying the request
      version:
        type: string
      result:
        type: object
        properties:
          success:
            type: boolean
          status:
            type: string
          content:
            $ref: "#/definitions/Client"
  ClientListResponse:
    title: List response
    type: object
    properties:
      id:
        type: string
        readOnly: true
        description: unique id identifying the request
      version:
        type: string
      result:
        type: object
        properties:
          success:
            type: boolean
          status:
            type: string
          metadata:
            $ref: "#/definitions/ResponseMetadata"
          content:
            type: array
            items:
              $ref: "#/definitions/Client"
...